+++ spec/fixtures/with_rb_extension/spec/foo.rb
@@ -1,0 +1,2 @@
+require 'rails_helper.rb'
+require 'spec_helper.rb'